Do not use an AST to calculate foldings (and refactor java foldings)
- Top level classes are now folded
- Only 1 folding region may start in a line (avoids problems when
un/collapsing via the +/- buttons)
- Extended folding is now only applied as when an element is known to
have children (because it is an IParent) but its children are unknown
(parent.getChildren() returns an empty array)
- Decouple the logic used for custom folding from the logic used for
standard and extended folding, improving code readability and
maintainability and also providing custom folding functionality to
classes that override the protected method
computeFoldingStructure(IJavaElement element,
FoldingStructureComputationContext ctx)
- Cases inside of switch expressions and switch statements are not
folded anymore

Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/848
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2085
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2439
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2454
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2470
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2571
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2596
Fixes https://github.com/eclipse-jdt/eclipse.jdt.ui/issues/2760

+++ org.eclipse.jdt.ui.tests/ui/org/eclipse/jdt/ui/tests/core/rules/ProjectTestSetup.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 IBM Corporation and others.
+ * Copyright (c) 2020, 2026 IBM Corporation and others.
  *
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
@@ -99,7 +99,7 @@ public class ProjectTestSetup extends ExternalResource {
 
 	@Override
 	protected void after() {
-		if (fJProject != null) {
+		if (fJProject != null && fJProject.exists()) {
 			try {
 				JavaProjectHelper.delete(fJProject);
 				CoreUtility.setAutoBuilding(fAutobuilding);
